febtools: support functions tested for the jintegral
The j-integral, as of now, does not match G for the
isotropic elastic case.  The value of J is very small,
so I suspect something is wrong with the material
definition or the J integral definition.

The integration function does have a test case, so it is probably
working.

The F tensor in the Element class does not have a test case,
and may be broken.