Adjust hex8 f tensor test to use twisted bar
The use of the twisted bar ensures that every tensor component is
tested.  Also, all the elements are included in the test, save for
the rigid body.

The rigidi body does have a discrepancy between the calculated F
tensor and the one stored in the log file.  I am not sure why this
is the case, nor am I sure if this is a problem.