Implement "easy" cyclic BCs and control sections
This is meant as part of a move to a pattern in which one function
generates the boundary condition curve, a second generates the
appropriate control section for FEBio, and a third combines them into a
single solution step.  With further development, the third function may
be subsumed into a "SolutionStep" class.