Support the special format for variable_data data
The special format of the variable_data block's data is: id, len, data

The id and len values are useless, so we used to just strip them.  But
it turns out that in FEBio 2.8, and probably as early as 2.6, element
variables [1] get one special format block per domain/region.  Junk id and
len values are thus sprinkled in the middle of the data.

Therefore this commit adds support for explicitly iterating over these
special sub-blocks in variable_data's data.

[1] The FEBio binary database specification calls by-element variables
"domain variables".