Bugfix: Don't flatten unpacked data of length 1
Sets of nodes, facets, and/or elements of length 1 do exist.  These will
have associated variable data with 1 value.  For consistency,
`unpack_variable_data` must return the unpacked variable data as a
tuple, like it does for larger sets of geometric items.

For data that is supposed to be a bare float or integer when
unpacked (like IDs and type/format codes), the caller must flatten the
unpacked data.