Retain material labels when writing to XML
Material labels should have already been retained when converting a
model to XML.  This was not working because, in input.py,
FebReader.mesh() and FebReader.model() both called
FebReader.materials().  This caused the material dictionary stored in
model.materials to use a different set of material objects than the
materials assigned to the mesh's element objects.  This broke the
pairing of material objects with material labels in output.py:xml().

To allow use of a consistent set of material objects, a optional parameter
(dict: material ID → material object) was added to
input.py:FebReader.mesh() so that it can use existing material IDs and
material objects when creating its Element objects.