Preserve node set names through import & export
All node sets must be frozen sets so that they're hashable, or the
NameRegistry lookup will break.  So the cycle to edit a node set is
convert it to a regular set, edit it, and convert it back to a
frozenset.  This helps enforce discipline regarding modifying node sets,
which is good, but the edit cycle is non-obvious, which is bad.

The problem of handling named node sets will have to be revisited in the
future.