Fix storing of dict as RigidBody density
The constructor for RigidBody was defined such that if it was passed a
dictionary of properties and nothing else, the entire dictionary would be
assigned to the density field.

Now, only the density value is assigned to the density field.  (The
other properties are lost according to the current design.)