Revert "Write <Step> elements to top of FEBio XML"
This reverts commit 101581d698c865233c6867c057f47dd515e73aae.

Some files, including Harrah's
`hydration_updatedconeqn_withmultigen.feb`, fail when the <Step>
elements are at the beginning of the file.  This The FEBio user manual
says: "The multi-step analysis feature introduces a new section to the
input ﬁle. Each step requires its own step section, preferably at the
bottom of the input ﬁle."  "Preferably" implies that placement at the
bottom of the input file is not required, so I consider this to be a
bug, either in FEBio's documentation or in the FEBio program.
Regardless of the cause, I can't put <Step> at the beginning of the file
if it sometimes creates errors.