Make parse_blocks tolerant of malformed data
It is useful for the xplt parser to continue through data blocks that
appear malformed so that a full parse tree is produced and can be
inspected by a human for possible errors.  Sometimes data will be
flagged as "malformed" because febtools doesn't understand it; sometimes
it will be a genuinely corrupt file.  Regardless, it is useful to be
able to produce a parse tree.