Modify textdata_table to enforce sequential step IDs
FEBio, when writing text data files, restarts time step ID numbering for
each analysis step (each <Step> section).  This is inconvenient because
this leaves us with no unique integer ID with which to refer to specific
time steps in a text data file.  Therefore, textdata_table was modified
so that it returns step IDs that increase sequentially with time.