Remove XpltBlocks
`XpltBlocks` was an unfinished a class meant to provide an
object-oriented representation of FEBio plotfile blocks.  It is not used
in any external code, and I don't intend to develop it further because
plain functions and data structures are doing an adequate job for now.
So I removed it to reduce clutter.