Add an unpack function that looks up the block format
To make unpacking block data easier, add a function that takes (a) the
block data and (b) the block tag and does what is correct for that block
tag.  In particular, the new function `unpack_block` will return either
a tuple or a single value (int or float) if the block type is known to
always store a single value.  Asking the caller to input the block's
data format, whether it stores a single value, and the endianness would
have been too burdensome.