Refactor to avoid circular imports
Split modules up so basic concepts are in standalone modules (such as
core.py, element.py, exceptions.py, material.py, and util.py),
higher-level FEA concepts are in different modules (model.py, added in
this commit) which import the basic modules, and read/modify/write
operations are in a third set of modules that import the other two
module hierarchies.

This refactoring was motivated by the observation that core.py and
util.py were importing each other due to the use of `Sequence` in `util`
and the use of `find_closest_timestep` in `core.Model`, which caused
ImportError to be raised in some circumstances.

All tests pass.