Fix: don't assume cause of NotImplementedError
Before: Any `NotImplementedError` raised from `mat_obj_from_elemd` was
assumed to be caused by an unsupported material type.  But
`NotImplementedError` can also be raised for unsupported types of
material orientation.

Now: `NotImplementedError` is no longer captured.  Therefore all
materials in a file must be supported before that file can be read,
which is probably for the best anyway.