Store nodes in an array
Some code, such as `conditions.prescribe_deformation`, assumes that the
mesh nodes data structure can be indexed into using an array of
integers.  This only works if the nodes data structure is an array.
Rather than doing a lot of conversion, just make the nodes data
structure an array (with the first index across nodes, like the prior
list of tuples data structure).

An array is more memory-efficient anyway.