Test intra-element F tensor heterogeneity
FEBio evaluates element data values at the element's Gauss points.  The
values in the logfile and plotfile output are averages over the Gauss
points.  This averaging is not readily apparent in the F tensor output,
but is very obvious for strain and stress tensors.  Therefore, check
that averaging Cauchy stress over an element's Gauss points in febtools
produces the same stress value as FEBio.

It's unclear if averaging stress this way is more accurate than just
evaluating stress at the center of the element.

All tests confirmed to pass.