Fix step → global time adjustment in FEBio XML
Before this commit, sequences that were used to define step-specific
prescribed nodal boundary conditions were not having their time values
adjusted to translate from step time to global time when written to
FEBio XML.  The code that looked for sequences to adjust was skipping
some sequences because its was checking an obsolete "variable" flag.
This commit fixes the comparison for both Sequence and ScaledSequence
sequences.