Indicate febtools can be type checked
Febtools is still mostly untyped; this is the very beginning of what
will be a long process to take advantage of static type checking.

Run `mypy --config-file febtools/mypy.ini febtools` from one directory
above the febtools package.  mypy isn't meant to be executed when the
current directory is within a package.  Unfortunately, this means mypy
will not pick up the `mypy.ini` file in febtools, so `--config-file
/path/to/mypy.ini` is needed.