Write <Surface> and <SurfacePair> once per name
Prior to this commit, `contact_section` wrote <Surface> and
<SurfacePair> elements into <Geometry> as needed.  Since
`contact_section` is called once for global contacts and then again once
per simulation step, this could result in writing multiple XML elements
for the same face set (surface) or surface pair.

Now, the <Surface> and <SurfacePair> elements are written all at once as
the last step in writing the FEBio XML tree.  Name registries are used
to keep track of (autogenerated) names in the meantime.