Don't treat warnings as errors in tests
Turning warnings into errors makes pytest terminate in an error if a
warning is raised during test collection.  And because the warning has
been transmuted into an error, pytest's options to filter warnings do
nothing.  This is a particular problem when the warning is raised in a
third-party module.

It is better to let pytest (or whatever test runner is being used)
handle detection and reporting of warnings rather than transmute them
into errors.