Make EntitySet objects safe for use as keys
In Python, hashable objects that compare equal must have the same hash
value.  Prior to this commit, two EntitySet objects a and b would
compare equal if they contained the same entities (IDs).  This commit
restores the mandatory invariant by making a == b only if a and b refer
to the same object (a is b).

This means that, from now on, two EntitySets must be compared for
equality of values using their symmetric difference (^), not eq (==).
Which is a little surprising and therefore not ideal.