Make (named) entity sets immutable
It turns out that detecting whether a name needs to be autogenerated for
a particular set of nodes / faces / elements / etc. is a lot easier if
existing entity sets are immutable and thus can be looked up by value in
the name registry (i.e., a hash table).  Mutable entity sets aren't
useful for much anyway and are not really safe to share between models.