Fix: check FEBio XML parameter type
Before this commit, the type of `f`, which is a dataclass field, was
being checked to determine if the named parameter was optional or
required.  But the type of a dataclass field, of course, is always
Field.  We instead want to check the FEBio XML rules for the parameter,
which come from the febioxml* modules.