Handle density material property
FEBio XML stores a variety of material properties in a material's tag,
including density.  Currently, material objects are instantiated using
the material class's `from_feb` method, which does not have a density
keyword argument.  Therefore we need to avoid passing "density" to the
`from_feb` method.  In this commit, this is accomplished by dividing the
XML element's properties into constitutive properties, which will be
passed to `from_feb`, and everything else.