Update fixtures to FEBio XML 4 & fix rigid rotations
The update to the fixtures was needed because FEBio 4.3 cannot read
Control/max_ups in FEBio XML 2.5, and FEBio 4.3 is a heavily used
release in my work.

That in turn required support for reading and writing the center of mass
of FEBio rigid materials, or the prescribed rigid rotations would be
different than intended.  FEBio applies rigid rotations about the
defined center of mass.

Rigid bodies are conceptually poorly defined so this led to cascading
changes and a giant commit.